terence's review
